Issue #586: Only build the non-root checks on !windows
This fixes
pkg/crc/preflight/preflight_checks_common.go:73:6: `checkIfRunningAsNormalUser` is unused (deadcode)
func checkIfRunningAsNormalUser() (bool, error) {
     ^

pkg/crc/preflight/preflight_checks_common.go:81:6: `fixRunAsNormalUser` is unused (deadcode)
     func fixRunAsNormalUser() (bool, error) {
          ^

https://github.com/code-ready/crc/issues/586
