+++ pkg/crc/preflight/preflight_checks_common.go
@@ -69,15 +69,3 @@ func fixOcBinaryCached() (bool, error) {
 	logging.Debug("oc binary cached")
 	return true, nil
 }
-
-func checkIfRunningAsNormalUser() (bool, error) {
-	if os.Geteuid() != 0 {
-		return true, nil
-	}
-	logging.Debug("Ran as root")
-	return false, fmt.Errorf("crc should be ran as a normal user")
-}
-
-func fixRunAsNormalUser() (bool, error) {
-	return false, fmt.Errorf("crc should be ran as a normal user")
-}
